requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
openai
huggingface-hub
sentence-transformers
pypdf
streamlit
tiktoken
python-dotenv
langchain-huggingface
transformers
torch
